pyproject.toml
[tool.poetry]
name = "pixida"
version = "0.1.0"
description = ""
authors = ["thiagodias"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.21.3"
structlog = "^22.1.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^5.0.4"
django-extensions = "^3.2.0"
ipython = "^8.4.0"
black = "^22.6.0"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
